+++ embedded_darwin_amd64.go
@@ -1,0 +1,8 @@
+//go:build darwin && amd64
+
+package main
+
+import "embed"
+
+//go:embed binaries/darwin-amd64/*
+var embeddedBinariesFS embed.FS