+++ embedded_darwin_arm64.go
@@ -1,0 +1,8 @@
+//go:build darwin && arm64
+
+package main
+
+import "embed"
+
+//go:embed binaries/darwin-arm64/*
+var embeddedBinariesFS embed.FS