+++ embedded_linux_arm64.go
@@ -1,0 +1,8 @@
+//go:build linux && arm64
+
+package main
+
+import "embed"
+
+//go:embed binaries/linux-arm64/*
+var embeddedBinariesFS embed.FS